Centralize recovery classifiers
