+++ server/src/services/recovery/index.ts
@@ -1,0 +1,40 @@
+export {
+  RECOVERY_KEY_PREFIXES,
+  RECOVERY_ORIGIN_KINDS,
+  RECOVERY_REASON_KINDS,
+  buildIssueGraphLivenessIncidentKey,
+  buildIssueGraphLivenessLeafKey,
+  parseIssueGraphLivenessIncidentKey,
+} from "./origins.js";
+export type {
+  RecoveryKeyPrefix,
+  RecoveryOriginKind,
+  RecoveryReasonKind,
+} from "./origins.js";
+export {
+  classifyIssueGraphLiveness,
+} from "./issue-graph-liveness.js";
+export type {
+  IssueGraphLivenessInput,
+  IssueLivenessAgentInput,
+  IssueLivenessDependencyPathEntry,
+  IssueLivenessExecutionPathInput,
+  IssueLivenessFinding,
+  IssueLivenessIssueInput,
+  IssueLivenessOwnerCandidate,
+  IssueLivenessOwnerCandidateReason,
+  IssueLivenessRelationInput,
+  IssueLivenessSeverity,
+  IssueLivenessState,
+} from "./issue-graph-liveness.js";
+export {
+  DEFAULT_MAX_LIVENESS_CONTINUATION_ATTEMPTS,
+  RUN_LIVENESS_CONTINUATION_REASON,
+  buildRunLivenessContinuationIdempotencyKey,
+  decideRunLivenessContinuation,
+  findExistingRunLivenessContinuationWake,
+  readContinuationAttempt,
+} from "./run-liveness-continuations.js";
+export type {
+  RunContinuationDecision,
+} from "./run-liveness-continuations.js";
