+++ server/src/services/recovery/origins.ts
@@ -1,0 +1,56 @@
+export const RECOVERY_ORIGIN_KINDS = {
+  issueGraphLivenessEscalation: "harness_liveness_escalation",
+  strandedIssueRecovery: "stranded_issue_recovery",
+  staleActiveRunEvaluation: "stale_active_run_evaluation",
+} as const;
+
+export const RECOVERY_REASON_KINDS = {
+  runLivenessContinuation: "run_liveness_continuation",
+} as const;
+
+export const RECOVERY_KEY_PREFIXES = {
+  issueGraphLivenessIncident: "harness_liveness",
+  issueGraphLivenessLeaf: "harness_liveness_leaf",
+} as const;
+
+export type RecoveryOriginKind = typeof RECOVERY_ORIGIN_KINDS[keyof typeof RECOVERY_ORIGIN_KINDS];
+export type RecoveryReasonKind = typeof RECOVERY_REASON_KINDS[keyof typeof RECOVERY_REASON_KINDS];
+export type RecoveryKeyPrefix = typeof RECOVERY_KEY_PREFIXES[keyof typeof RECOVERY_KEY_PREFIXES];
+
+export function buildIssueGraphLivenessIncidentKey(input: {
+  companyId: string;
+  issueId: string;
+  state: string;
+  blockerIssueId?: string | null;
+  participantAgentId?: string | null;
+}) {
+  return [
+    RECOVERY_KEY_PREFIXES.issueGraphLivenessIncident,
+    input.companyId,
+    input.issueId,
+    input.state,
+    input.blockerIssueId ?? input.participantAgentId ?? "none",
+  ].join(":");
+}
+
+export function parseIssueGraphLivenessIncidentKey(incidentKey: string | null | undefined) {
+  if (!incidentKey) return null;
+  const parts = incidentKey.split(":");
+  if (parts.length !== 5 || parts[0] !== RECOVERY_KEY_PREFIXES.issueGraphLivenessIncident) return null;
+  const [, companyId, issueId, state, leafIssueId] = parts;
+  if (!companyId || !issueId || !state || !leafIssueId) return null;
+  return { companyId, issueId, state, leafIssueId };
+}
+
+export function buildIssueGraphLivenessLeafKey(input: {
+  companyId: string;
+  state: string;
+  leafIssueId: string;
+}) {
+  return [
+    RECOVERY_KEY_PREFIXES.issueGraphLivenessLeaf,
+    input.companyId,
+    input.state,
+    input.leafIssueId,
+  ].join(":");
+}
